Worked on authenticating pages for logged-in users.
